+++ lib/src/utils/common_currencies.dart
@@ -1,0 +1,34 @@
+/// List of the 30 most commonly used currencies in the world
+/// for physical money (bills and cards), excluding cryptocurrencies
+const List<String> commonCurrencyCodes = [
+  'USD', // United States Dollar
+  'EUR', // Euro
+  'GBP', // British Pound Sterling
+  'JPY', // Japanese Yen
+  'CNY', // Chinese Yuan
+  'AUD', // Australian Dollar
+  'CAD', // Canadian Dollar
+  'CHF', // Swiss Franc
+  'INR', // Indian Rupee
+  'MXN', // Mexican Peso
+  'BRL', // Brazilian Real
+  'ZAR', // South African Rand
+  'RUB', // Russian Ruble
+  'KRW', // South Korean Won
+  'SGD', // Singapore Dollar
+  'HKD', // Hong Kong Dollar
+  'NOK', // Norwegian Krone
+  'SEK', // Swedish Krona
+  'DKK', // Danish Krone
+  'PLN', // Polish Zloty
+  'ARS', // Argentine Peso
+  'CLP', // Chilean Peso
+  'COP', // Colombian Peso
+  'THB', // Thai Baht
+  'MYR', // Malaysian Ringgit
+  'IDR', // Indonesian Rupiah
+  'NZD', // New Zealand Dollar
+  'TRY', // Turkish Lira
+  'SAR', // Saudi Riyal
+  'AED', // UAE Dirham
+];